refactor: 認証系E2Eヘルパーをshared配下に集約 HOXBL-87
- auth/helpers/test-setup.ts を shared/helpers/auth-session.ts へ移設
- auth・auth-emailの各specから新しいパスを参照するよう更新
- todo等の他featureからも同一ヘルパーを一貫して再利用できるようにする
