+++ app/client/e2e/shared/helpers/auth-session.ts
